Update how-to-host-a-wcf-service-in-a-managed-application.md
Article use too often term 'Managed Application'. The meaning of this term is not really clear.  It was used almost two decades ago to express application host build on top of managed ,NET framework in contrast to pure C++, MFC or whatever has been used those days known as "unmanaged".
It is in this article simply wrong name.
Propose: Host a WCF Service in a Console Application.